use float instead of np.float64

+++ ipi/inputs/initializer.py
@@ -99,7 +99,7 @@ class InputInitBase(InputValue):
         if self.mode.fetch() == "manual":
             if "[" in value and "]" in value:  # value appears to be a list
                 if self._storageclass is float:
-                    value = io_xml.read_array(np.float64, value)
+                    value = io_xml.read_array(float, value)
                 else:
                     value = io_xml.read_list(value)
             else:
@@ -321,7 +321,7 @@ class InputInitCell(InputInitBase):
         if mode == "abc" or mode == "abcABC":
 
             h = io_xml.read_array(
-                np.float64, ibase.value
+                float, ibase.value
             )  # As of numpy v1.20, numpy.float as well as similar aliases (including numpy.int) were deprecated
 
             if mode == "abc":
